// src/store/actions/rooms/interfaces.ts
export enum getRoomsType {
  GET_ROOMS = 'GET_ROOMS',
  GET_ROOMS_FAIL = 'GET_ROOMS_FAIL',
  GET_ROOMS_SUCCESS = 'GET_ROOMS_SUCCESS',
  SHOW_ROOM = 'SHOW_ROOM',
  RESET_EDITING = 'RESET_EDITING',
  SHOW_WARNING = 'SHOW_WARNING',
  CANCEL_DELETE = 'CANCEL_DELETE',
  GET_URL = 'GET_URL',
  CANCEL_UPLOAD = 'CANCEL_UPLOAD',
  EDIT_ROOM_FAIL = 'EDIT_ROOM_FAIL',
  EDIT_ROOM_SUCCESS = 'EDIT_ROOM_SUCCESS',
  ADD_ROOM_FAIL = 'ADD_ROOM_FAIL',
  ADD_ROOM_SUCCESS = 'ADD_ROOM_SUCCESS',
}